Basic Announcements Edit and Viewing Fragments
Basic UI view and backend for loading announcements for the user.

Basic UI view and backend for organizer being able to edit, delete, and add announcements.

All connected to the database but key implementations are missing such as how to get the event ID from the the joined event or sending push notifications.